fix: set line number type to u64 and give option to propagate into tail_log_files

+++ src/event_models.rs
@@ -33,7 +33,7 @@ pub struct Entity {
 
 #[derive(Debug, Clone, Default)]
 pub struct CombatEvent {
-    pub line_number: usize,
+    pub line_number: u64,
     pub timestamp: Timestamp,
     pub source_entity: Entity,
     pub target_entity: Entity,
